\name{BayClone2_Simulation1_tot}
\docType{data}
\alias{BayClone2_Simulation1_tot}
\title{Total number of reads (N) for simulation 1 in Lee et al (2014)}
\description{
This is a (S*T) data matrix for simulation 1 with S=100 and T=4 in Lee, et al.  Each element of this data matrix gives the total number of reads mapped at locus s in sample t.
}
\usage{BayClone2_Simulation1_tot}
\format{A integer matrix with 100 rows and 4 columns.}
\references{
    J. Lee, P. Mueller, S. Sengupta, K. Gulukota, Y. Ji, Bayesian Inference for Tumor Subclones Accounting for Sequencing and Structural Variants (http://arxiv.org/abs/1409.7158)
}
\keyword{datasets}